Move ProcessState and ProcessStepState into ProcessManager module
