+++ apps/dh/api-dh/source/DataHub.WebApi/Modules/ProcessManager/Calculations/CalculationsNode.cs
@@ -18,11 +18,11 @@ using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS
 using Energinet.DataHub.WebApi.Clients.MarketParticipant.v1;
 using Energinet.DataHub.WebApi.Modules.Common;
 using Energinet.DataHub.WebApi.Modules.Common.DataLoaders;
-using Energinet.DataHub.WebApi.Modules.Common.Models;
 using Energinet.DataHub.WebApi.Modules.ProcessManager.Calculations.Client;
 using Energinet.DataHub.WebApi.Modules.ProcessManager.Calculations.Enums;
 using Energinet.DataHub.WebApi.Modules.ProcessManager.Calculations.Types;
 using Energinet.DataHub.WebApi.Modules.ProcessManager.Orchestrations.Types;
+using Energinet.DataHub.WebApi.Modules.ProcessManager.Types;
 using HotChocolate.Authorization;
 using HotChocolate.Subscriptions;
 using NodaTime;
